requirements.txt
albumentations==1.3.0
anyio==3.6.2
anykeystore==0.2
apex==0.9.10.dev0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens 
attrs==22.2.0
backcall 
backports.functools-lru-cache
beautifulsoup4==4.11.1
bleach==6.0.0
brotlipy
certifi==2022.12.7
cffi 
chardet==3.0.4
charset-normalizer 
Click==7.0
comm 
contourpy==1.0.6
cryptacular==1.6.2
cryptography 
cycler==0.11.0
debugpy 
decorator 
defusedxml==0.7.1
entrypoints
executing
fastjsonschema==2.16.2
filelock==3.8.2
fonttools==4.38.0
fqdn==1.5.1
greenlet==2.0.1
huggingface-hub==0.11.1
hupper==1.10.3
idna 
imageio==2.22.4
ipykernel
ipython 
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi 
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core 
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lab==7.2
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline
mistune==2.0.4


natsort==8.2.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio 
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging 
pandas==1.5.2
pandocfilters==1.5.0
parso 
PasteDeploy==3.0.1
pbkdf2==1.3
pexpect 
pickleshare 
Pillow==9.3.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs 
prometheus-client==0.16.0
prompt-toolkit
protobuf==3.20.1
psutil 
ptyprocess 
pure-eval
pycparser
Pygments 
pyOpenSSL 
pyparsing==3.0.9
pyramid==2.0
pyramid-mailer==0.15.1
pyrsistent==0.19.3
PySocks 
python-dateutil 
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
repoze.sendmail==4.4.1
requests
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
Send2Trash==1.8.0
simplejson==3.18.1
six 
sniffio==1.3.0
sotabenchapi==0.0.16
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
stack-data
tabulate==0.8.3
tensorboardX==2.5.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
tinycss2==1.2.1

torchbench==0.0.31
torchsummary==1.5.1

tornado==6.2
tqdm==4.64.1
traitlets 
transaction==3.0.1
translationstring==1.4
txt2tags==3.7
typing_extensions
uri-template==1.2.0
urllib3 
velruse==1.1.1
venusian==3.0.0
wcwidth 
webcolors==1.12
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.5.0
widgetsnbextension==4.0.5
WTForms==3.0.1
wtforms-recaptcha==0.3.2
yacs==0.1.8
zope.deprecation==4.4.0
zope.interface==5.5.2
zope.sqlalchemy==1.6
